mejora de repositorios y servicios

+++ src/main/java/com/gamm/vinos_api/domain/model/Sucursal.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -14,7 +16,9 @@ public class Sucursal {
   private Integer idSucursal;
   private String nombre;
   private String ubicacion;
-  private Double latitud;
-  private Double longitud;
+  private BigDecimal  latitud;
+  private BigDecimal longitud;
   private EstadoSucursal estado;
+
+
 }
